Add super field to blocks
Adds a new `super` field to blocks that lists the (obfuscated) names of the super classes for the block class. I added this so I can iterate over the class + super classes to find the class that has the field_name in the mappings.

I tested this for 1.8, 1.12, 1.13, and 1.19 and it works for all of them.